+++ platform/ios/src/Mapbox.h
@@ -14,6 +14,7 @@ FOUNDATION_EXPORT MGL_EXPORT const unsigned char MapboxVersionString[];
 #import "MGLAnnotationImage.h"
 #import "MGLCalloutView.h"
 #import "MGLClockDirectionFormatter.h"
+#import "MGLCluster.h"
 #import "MGLCompassDirectionFormatter.h"
 #import "MGLCoordinateFormatter.h"
 #import "MGLDistanceFormatter.h"