+++ platform/macos/src/Mapbox.h
@@ -12,6 +12,7 @@ FOUNDATION_EXPORT MGL_EXPORT const unsigned char MapboxVersionString[];
 #import "MGLAnnotation.h"
 #import "MGLAnnotationImage.h"
 #import "MGLClockDirectionFormatter.h"
+#import "MGLCluster.h"
 #import "MGLCompassDirectionFormatter.h"
 #import "MGLCoordinateFormatter.h"
 #import "MGLDistanceFormatter.h"